// src/web/shared/Text/types.ts
export interface StyledTextProps {
  color?: string;
  align?: string;
  size?: string;
  weight?: string;
  font?: string;
}

export interface TextProps {
  color?: string;
  align?: string;
  size?: string;
  weight?: string;
  font?: string;
  children?: any;
  component?: string;
  className?: string;
}